fix: introduce _pydantic.py for pydantic basic reexports
