+++ .flake8
@@ -12,4 +12,5 @@ exclude =
     .nox,
     .pytest_cache
     __pycache__,
+    _pydantic.py,
     __init__.py
