+++ aidial_sdk/chat_completion/form.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 from typing import (
-    TYPE_CHECKING,
     Any,
     Callable,
     Dict,
@@ -19,28 +18,14 @@ from typing import (
 
 from pydantic.v1.validators import make_literal_validator
 
-from aidial_sdk.pydantic._compat import PYDANTIC_V2
+from aidial_sdk._pydantic import (
+    PYDANTIC_V2,
+    BaseModel,
+    FieldInfo,
+    ModelMetaclass,
+    validator,
+)
 from aidial_sdk.pydantic._model_config import ModelConfigWrapper
-
-if TYPE_CHECKING:
-    from pydantic import BaseModel
-    from pydantic import field_validator as validator
-    from pydantic._internal._model_construction import ModelMetaclass
-    from pydantic.fields import FieldInfo
-else:
-    if PYDANTIC_V2:
-        from pydantic import BaseModel
-        from pydantic import field_validator as validator
-        from pydantic._internal._model_construction import ModelMetaclass
-        from pydantic.fields import FieldInfo
-    else:
-        from pydantic.v1 import BaseModel, validator
-
-        try:
-            from pydantic.v1.main import ModelMetaclass
-        except ImportError:
-            from pydantic.main import ModelMetaclass
-        from pydantic.v1.fields import FieldInfo
 
 _T = TypeVar("_T")
