--- aidial_sdk/pydantic/_version.py
@@ -1,3 +1,0 @@
-import pydantic
-
-INSTALLED_PYDANTIC_V2 = pydantic.VERSION.startswith("2.")
