+++ aidial_sdk/utils/pydantic.py
@@ -1,4 +1,5 @@
-from aidial_sdk.pydantic._compat import PYDANTIC_V2, BaseModel, ConfigDict
+from aidial_sdk._pydantic import PYDANTIC_V2, ConfigDict
+from aidial_sdk.pydantic._compat import BaseModel
 
 
 class ExtraForbidModel(BaseModel):
